refactor: change some naming
